requirements.txt
# AstrBot 洛谷助手插件依赖

# 网页自动化
playwright>=1.40.0

# 图表生成
matplotlib>=3.7.0

# 验证码识别（可选）
ddddocr>=1.4.0

# HTTP 请求
requests>=2.28.0

# 数据处理
pillow>=9.0.0
